Expose pretty-dom utility
