+++ src/index.js
@@ -11,3 +11,4 @@ export * from './matches'
 export * from './get-node-text'
 export * from './events'
 export * from './bind-element-to-queries'
+export * from './pretty-dom'
