+++ src/queries.js
@@ -1,8 +1,6 @@
-import prettyFormat from 'pretty-format'
 import {matches} from './matches'
 import {getNodeText} from './get-node-text'
-
-const {DOMElement, DOMCollection} = prettyFormat.plugins
+import {prettyDOM} from './pretty-dom'
 
 // Here are the queries for the library.
 // The queries here should only be things that are accessible to both users who are using a screen reader
@@ -74,7 +72,7 @@ function getByTestId(container, id, ...rest) {
   const el = queryByTestId(container, id, ...rest)
   if (!el) {
     throw new Error(
-      `Unable to find an element by: [data-testid="${id}"] \n\n${htmlElementToDisplay(
+      `Unable to find an element by: [data-testid="${id}"] \n\n${prettyDOM(
         container,
       )}`,
     )
@@ -86,7 +84,7 @@ function getByPlaceholderText(container, text, ...rest) {
   const el = queryByPlaceholderText(container, text, ...rest)
   if (!el) {
     throw new Error(
-      `Unable to find an element with the placeholder text of: ${text} \n\n${htmlElementToDisplay(
+      `Unable to find an element with the placeholder text of: ${text} \n\n${prettyDOM(
         container,
       )}`,
     )
@@ -100,13 +98,13 @@ function getByLabelText(container, text, ...rest) {
     const label = queryLabelByText(container, text)
     if (label) {
       throw new Error(
-        `Found a label with the text of: ${text}, however no form control was found associated to that label. Make sure you're using the "for" attribute or "aria-labelledby" attribute correctly. \n\n${htmlElementToDisplay(
+        `Found a label with the text of: ${text}, however no form control was found associated to that label. Make sure you're using the "for" attribute or "aria-labelledby" attribute correctly. \n\n${prettyDOM(
           container,
         )}`,
       )
     } else {
       throw new Error(
-        `Unable to find a label with the text of: ${text} \n\n${htmlElementToDisplay(
+        `Unable to find a label with the text of: ${text} \n\n${prettyDOM(
           container,
         )}`,
       )
@@ -119,7 +117,7 @@ function getByText(container, text, ...rest) {
   const el = queryByText(container, text, ...rest)
   if (!el) {
     throw new Error(
-      `Unable to find an element with the text: ${text}. This could be because the text is broken up by multiple elements. In this case, you can provide a function for your text matcher to make your matcher more flexible. \n\n${htmlElementToDisplay(
+      `Unable to find an element with the text: ${text}. This could be because the text is broken up by multiple elements. In this case, you can provide a function for your text matcher to make your matcher more flexible. \n\n${prettyDOM(
         container,
       )}`,
     )
@@ -139,24 +137,12 @@ function getByAltText(container, alt) {
   const el = queryByAltText(container, alt)
   if (!el) {
     throw new Error(
-      `Unable to find an element with the alt text: ${alt} \n\n${htmlElementToDisplay(
+      `Unable to find an element with the alt text: ${alt} \n\n${prettyDOM(
         container,
       )}`,
     )
   }
   return el
-}
-
-function htmlElementToDisplay(htmlElement) {
-  const debugContent = prettyFormat(htmlElement, {
-    plugins: [DOMElement, DOMCollection],
-    printFunctionName: false,
-    highlight: true,
-  })
-  const maxLength = process.env.DEBUG_PRINT_LIMIT || 7000
-  return htmlElement.outerHTML.length > maxLength
-    ? `${debugContent.slice(0, maxLength)}...`
-    : debugContent
 }
 
 export {
